Added BLAS GEMM. Added transpose_op to GEMV. Back to <double> (instead of <float>) matrices. Proper vectors handling.

+++ c_src/pcuda_matrix_float_buffer.cpp
@@ -15,14 +15,14 @@
 #define IDX2CCM(idx,R) (idx/R) //index to col for CM matrix
 
 PCudaMatrixFloatBuffer::PCudaMatrixFloatBuffer() {
-    this->data = new std::vector<float>();
+    this->data = new std::vector<double>();
     this->_rows = 1;
 }
 
 PCudaMatrixFloatBuffer::PCudaMatrixFloatBuffer(unsigned  rows, unsigned  cols) {
     this->_rows = rows;
     this->_cols = cols;
-    this->data = new std::vector<float>(rows*cols);
+    this->data = new std::vector<double>(rows*cols);
 }
 
 PCudaMatrixFloatBuffer::~PCudaMatrixFloatBuffer() {
@@ -48,11 +48,11 @@ void PCudaMatrixFloatBuffer::write(ErlNifEnv *env, ERL_NIF_TERM data) {
       while (enif_get_list_cell(env, head_row, &head, &head_row))
         if (enif_get_double(env, head, &value)) {
             //this->data->push_back(value);
-            this->data->at((IDX2C(IDX2RRM(idx,C), IDX2CRM(idx,C), ld)) ) = (float)value;
+            this->data->at((IDX2C(IDX2RRM(idx,C), IDX2CRM(idx,C), ld)) ) = value;
             ++idx;
         }else if (enif_get_long(env, head, &lvalue)) {
             //this->data->push_back((double)lvalue);
-            this->data->at(IDX2C(IDX2RRM(idx,C), IDX2CRM(idx,C), ld)) = (float)lvalue;
+            this->data->at(IDX2C(IDX2RRM(idx,C), IDX2CRM(idx,C), ld)) = (double)lvalue;
             ++idx;
         }
     
@@ -64,7 +64,7 @@ void PCudaMatrixFloatBuffer::write(ErlNifEnv *env, ERL_NIF_TERM data) {
 
 //converts column major vector to Erlang row major "list-of-lists"
 ERL_NIF_TERM PCudaMatrixFloatBuffer::toErlTerms(ErlNifEnv *env) {
-    std::vector<float>::iterator iter;
+    std::vector<double>::iterator iter;
     ERL_NIF_TERM retval = enif_make_list(env, 0);
     ERL_NIF_TERM row;
